requirements.txt
presto-python-client
psycopg2-binary
